Move to_date Function
move to_date to utils.py
fixed function call
